2024-3-28 1、新增soda账号退出 2、优化移动端导航栏显示

+++ src/store/index.js
@@ -88,11 +88,18 @@ export default new Vuex.Store({
     },
 
     // 退出登录
-    logout (state) {
-      this.commit('clearCookies')
-      window.sessionStorage.clear()
-      state.thirdParty.userInfo = { avatarUrl: '' }
-      state.thirdParty.isLogin = false
+    logout (state, platform) {
+      if (platform === 'sodamusic') {
+        window.localStorage.clear()
+        state.sodaAccount.userInfo = {}
+        state.sodaAccount.isLogin = false
+      }
+      if (platform === 'neteasecloud') {
+        this.commit('clearCookies')
+        window.sessionStorage.clear()
+        state.thirdParty.userInfo = { avatarUrl: '' }
+        state.thirdParty.isLogin = false
+      }
     },
 
     // 清除本地cookie
